% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice6-package.R
\docType{package}
\name{practice6-package}
\alias{practice6}
\alias{practice6-package}
\title{practice6: Practice 6 - R Package}
\description{
Contains the structure of an R package, result of the practice 6 of the workshop 'Data Toolbox for Reproducible Research in Ecology' \url{https://rdatatoolbox.github.io/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rdatatoolbox/practice6}
  \item Report bugs at \url{https://github.com/rdatatoolbox/practice6/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
